requirements.txt
certifi==2020.12.5
Cython==0.29.22
numpy==1.19.5
pandas==1.1.5
Pillow==8.2.0
plyfile==0.7.3
protobuf==3.15.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
scipy==1.5.4
six==1.15.0
tensorboardX==2.1
torch==1.3.1
torchvision==0.4.2
tqdm==4.59.0
trimesh==3.9.10
